requirements.txt
numba
numpy
opencv-python
pillow
